+++ src/app/layout/loading-indicator.service.ts
@@ -1,0 +1,27 @@
+import { Injectable } from '@angular/core';
+import {
+  NavigationCancel,
+  NavigationEnd,
+  NavigationError,
+  NavigationStart,
+  Router
+} from '@angular/router';
+import { Observable } from 'rxjs';
+import { filter, map } from 'rxjs/operators';
+
+@Injectable({ providedIn: 'root' })
+export class LoadingIndicatorService {
+  public isLoading$: Observable<boolean>;
+  constructor(private router: Router) {
+    this.isLoading$ = this.router.events.pipe(
+      filter(
+        event =>
+          event instanceof NavigationStart ||
+          event instanceof NavigationEnd ||
+          event instanceof NavigationCancel ||
+          event instanceof NavigationError
+      ),
+      map(event => !!(event instanceof NavigationStart))
+    );
+  }
+}